refactor: #42 사용자 도메인 리펙터링
- classroomappender.addChild -> classroom.addChild로 변경
- phoneNumber 미사용 함수 제거
