+++ src/main/kotlin/com/asap/asapbackend/domain/classroom/domain/model/Classroom.kt
@@ -1,6 +1,8 @@
 package com.asap.asapbackend.domain.classroom.domain.model
 
+import com.asap.asapbackend.domain.child.domain.model.Child
 import com.asap.asapbackend.domain.school.domain.model.School
+import com.asap.asapbackend.domain.teacher.domain.model.Teacher
 import com.asap.asapbackend.global.domain.BaseDateEntity
 import jakarta.persistence.*
 import java.util.*
@@ -27,4 +29,19 @@ class Classroom(
 
     @ManyToOne(fetch = FetchType.LAZY)
     val school: School = school
+
+
+    @OneToMany(mappedBy = "classroom")
+    val childClassroomSet: MutableSet<ChildClassroom> = mutableSetOf()
+
+    @OneToMany(mappedBy = "classroom")
+    val teacherClassroomSet: MutableSet<TeacherClassroom> = mutableSetOf()
+
+    fun addChild(child: Child) {
+        childClassroomSet.add(ChildClassroom(child, this))
+    }
+
+    fun addTeacher(teacher: Teacher) {
+        teacherClassroomSet.add(TeacherClassroom(teacher, this))
+    }
 }
