+++ src/main/kotlin/com/asap/asapbackend/domain/user/application/UserService.kt
@@ -1,13 +1,12 @@
 package com.asap.asapbackend.domain.user.application
 
 import com.asap.asapbackend.domain.child.domain.service.ChildAppender
-import com.asap.asapbackend.domain.classroom.domain.service.ClassroomAppender
 import com.asap.asapbackend.domain.classroom.domain.service.ClassroomReader
 import com.asap.asapbackend.domain.user.application.dto.CreateUser
 import com.asap.asapbackend.domain.user.domain.service.UserAppender
-import com.asap.asapbackend.global.jwt.Claims
-import com.asap.asapbackend.global.jwt.JwtProvider
-import com.asap.asapbackend.global.jwt.TokenExtractor
+import com.asap.asapbackend.global.jwt.vo.Claims
+import com.asap.asapbackend.global.jwt.util.JwtProvider
+import com.asap.asapbackend.global.jwt.util.TokenExtractor
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -18,7 +17,6 @@ class UserService(
     private val tokenExtractor: TokenExtractor,
     private val userAppender: UserAppender,
     private val classroomReader: ClassroomReader,
-    private val classroomAppender: ClassroomAppender,
     private val childAppender: ChildAppender,
 ) {
 
@@ -26,15 +24,16 @@ class UserService(
     fun createUser(request: CreateUser.Request): CreateUser.Response {
         val user = request.extractUser {
             val registrationToken = tokenExtractor.extractValue(it)
-            jwtProvider.extractRegistrationClaimsFromToken(registrationToken)
+            val claims: Claims.RegistrationClaims = jwtProvider.extractRegistrationClaimsFromToken(registrationToken)
+            return@extractUser claims
         }.apply { userAppender.appendUser(this) }
 
-        request.children.forEach {
-            val classroom = it.extractClassroom { grade, className, classCode, schoolId ->
-                    classroomReader.findByClassInfoAndSchoolId(grade, className, classCode,schoolId)
-            }
+        request.performActionOnChildren {
             val child = it.extractChild(user).apply { childAppender.appendChild(this) }
-            classroomAppender.addChild(classroom, child)
+            it.performActionOnClassroom { grade, className, classCode, schoolId ->
+                val classroom = classroomReader.findByClassInfoAndSchoolId(grade, className, classCode, schoolId)
+                classroom.addChild(child)
+            }
         }
         val userClaims = Claims.UserClaims(user.id)
         return CreateUser.Response(
