+++ src/main/kotlin/com/asap/asapbackend/domain/user/domain/model/PhoneNumber.kt
@@ -10,10 +10,6 @@ data class PhoneNumber(
         validate()
     }
 
-    fun update(number: String): PhoneNumber {
-        return PhoneNumber(number)
-    }
-
     fun validate(){
         validateNumber(this.number)
     }
